Move gcp cli file in the correct folder
