+++ tests/providers/gcp/gcp_cli.py
